added move year

+++ src/main/kotlin/com/osia/nota_maestro/dto/school/v1/SchoolDto.kt
@@ -14,8 +14,8 @@ class SchoolDto : BaseDto() {
     var actualYear: Int? = null
     var enabledTeacher: Boolean? = null
     var enabledStudent: Boolean? = null
-    var maxNote: Int? = null
-    var minNote: Int? = null
+    var maxNote: Double? = null
+    var minNote: Double? = null
     var toLose: Int? = null
     var recoveryType: String? = null
     var enabledFinalRecovery: Boolean? = null
