+++ src/main/kotlin/com/osia/nota_maestro/dto/school/v1/SchoolRequest.kt
@@ -15,8 +15,8 @@ class SchoolRequest {
     var actualYear: Int? = null
     var enabledTeacher: Boolean? = null
     var enabledStudent: Boolean? = null
-    var maxNote: Int? = null
-    var minNote: Int? = null
+    var maxNote: Double? = null
+    var minNote: Double? = null
     var toLose: Int? = null
     var recoveryType: String? = null
     var enabledFinalRecovery: Boolean? = null
