+++ src/main/kotlin/com/osia/nota_maestro/dto/user/v1/UserDto.kt
@@ -27,8 +27,8 @@ class UserDto : BaseDto() {
     var actualYear: Int? = null
     var enabledTeacher: Boolean? = null
     var enabledStudent: Boolean? = null
-    var maxNote: Int? = null
-    var minNote: Int? = null
+    var maxNote: Double? = null
+    var minNote: Double? = null
     var toLose: Int? = null
     var recoveryType: String? = null
     var enabledFinalRecovery: Boolean? = null
